backend/requirements/base.txt
fastapi==0.115.0
uvicorn==0.34.0
celery==5.4.0
celery[redis]==5.4.0
mysql-connector-python==8.4.0
python-dotenv==1.0.1
itsdangerous>=2.0
PyJWT

Authlib~=1.3.2
starlette~=0.38.6
pydantic~=2.10.4
pydantic-settings
httpx
